+++ CHANGES.rst
@@ -8,6 +8,9 @@ Added
 
 - Display TOS and policy URI on the consent list page. :pr:`102`
 - Admin token deletion :pr:`100` :pr:`101`
+- Revoked consents can be restored. :pr:`103`
+- Pre-consented clients are displayed in the user consent list,
+  and their consents can be revoked. :issue:`69` :pr:`103`
 
 Fixed
 *****